lecture_llm_api/pyproject.toml
[project]
name = "lecture-llm-api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "nikgilae", email = "nikitaambrozyak@gmail.com" }
]
requires-python = ">=3.12"
dependencies = [
    "gigachat>=0.1.43",
    "openai>=2.6.0",
    "pydantic-settings>=2.11.0",
    "python-dotenv>=1.1.1",
    "rich>=14.2.0",
]

[project.scripts]
lecture-llm-api = "lecture_llm_api:main"

[build-system]
requires = ["uv_build>=0.8.17,<0.9.0"]
build-backend = "uv_build"

[dependency-groups]
dev = [
    "ipykernel>=7.1.0",
    "pip>=25.3",
]
